FIX: cyclic deps просто вынес ответы в пакет domain/response
