+++ internal/domain/response/OK.go
@@ -1,23 +1,10 @@
-package api
+package response
 
 import (
 	"fmt"
-	"rediska/internal/util/resper"
-
 	"net"
+	"rediska/internal/util/resper"
 )
-
-func PONG(conn net.Conn) {
-	buf, err := resper.EncodeSimpleString("PONG")
-	if err != nil {
-		fmt.Println("failed to encode:", err)
-	}
-
-	_, err = conn.Write(buf.Bytes())
-	if err != nil {
-		fmt.Println("failed to write PONG to client")
-	}
-}
 
 func OK(conn net.Conn) {
 	op := "OK response"
